Extract screen culler base logic
