+++ src/core/ScreenCuller/src/culler-renderer.ts
@@ -1,7 +1,6 @@
 import * as THREE from "three";
 import { Component, Disposable, Event } from "../../../base-types";
 import { Components } from "../../Components";
-import { Disposer } from "../../Disposer";
 import { readPixelsAsync } from "./screen-culler-helper";
 
 export interface CullerRendererSettings {
@@ -12,11 +11,9 @@ export interface CullerRendererSettings {
 }
 
 /**
- * A renderer to determine an object's visibility on screen
+ * A base renderer to determine visibility on screen
  */
-export class CullerRenderer extends Component<
-  Map<string, THREE.InstancedMesh>
-> {
+export class CullerRenderer extends Component<THREE.WebGLRenderer> {
   /** {@link Disposable.onDisposed} */
   readonly onDisposed = new Event<string>();
 
@@ -25,10 +22,7 @@ export class CullerRenderer extends Component<
    * meshes that are currently visible, and the ones that were visible
    * just before but not anymore.
    */
-  readonly onViewUpdated = new Event<{
-    seen: Set<THREE.Mesh>;
-    unseen: Set<THREE.Mesh>;
-  }>();
+  readonly onViewUpdated = new Event<any>();
 
   /** {@link Component.enabled} */
   enabled = true;
@@ -45,34 +39,22 @@ export class CullerRenderer extends Component<
    */
   renderDebugFrame = false;
 
-  readonly renderer: THREE.WebGLRenderer;
-
-  private _updateInterval = 1000;
   private _width = 512;
   private _height = 512;
-  private _autoUpdate = true;
 
+  protected autoUpdate = true;
+  protected updateInterval = 1000;
+
+  protected readonly worker: Worker;
+
+  protected readonly renderer: THREE.WebGLRenderer;
   private readonly renderTarget: THREE.WebGLRenderTarget;
+  protected readonly scene = new THREE.Scene();
+
   private readonly bufferSize: number;
-  private readonly materialCache: Map<string, THREE.MeshBasicMaterial>;
-  private readonly worker: Worker;
-
-  private _meshColorMap = new Map<string, THREE.Mesh>();
-  private _colorMeshes = new Map<string, THREE.InstancedMesh>();
-  private _meshes = new Map<string, THREE.Mesh>();
-
-  private _currentVisibleMeshes = new Set<THREE.Mesh>();
-  private _recentlyHiddenMeshes = new Set<THREE.Mesh>();
-
-  private readonly _transparentMat = new THREE.MeshBasicMaterial({
-    transparent: true,
-    opacity: 0,
-  });
 
   private _colors = { r: 0, g: 0, b: 0, i: 0 };
 
-  // Alternative scene and meshes to make the visibility check
-  private readonly _scene = new THREE.Scene();
   private readonly _buffer: Uint8Array;
 
   constructor(components: Components, settings?: CullerRendererSettings) {
@@ -85,7 +67,6 @@ export class CullerRenderer extends Component<
     this.renderTarget = new THREE.WebGLRenderTarget(this._width, this._height);
     this.bufferSize = this._width * this._height * 4;
     this._buffer = new Uint8Array(this.bufferSize);
-    this.materialCache = new Map<string, THREE.MeshBasicMaterial>();
 
     const code = `
       addEventListener("message", (event) => {
@@ -104,114 +85,27 @@ export class CullerRenderer extends Component<
 
     const blob = new Blob([code], { type: "application/javascript" });
     this.worker = new Worker(URL.createObjectURL(blob));
-    this.worker.addEventListener("message", this.handleWorkerMessage);
-    if (this._autoUpdate) {
-      window.setInterval(this.updateVisibility, this._updateInterval);
-    }
   }
 
   /**
    * {@link Component.get}.
-   * @returns the map of internal meshes used to determine visibility.
    */
   get() {
-    return this._colorMeshes;
+    return this.renderer;
   }
 
   /** {@link Disposable.dispose} */
   async dispose() {
     this.enabled = false;
-    this._currentVisibleMeshes.clear();
-    this._recentlyHiddenMeshes.clear();
-    this._scene.children.length = 0;
+    for (const child of this.scene.children) {
+      child.removeFromParent();
+    }
     this.onViewUpdated.reset();
     this.worker.terminate();
     this.renderer.dispose();
     this.renderTarget.dispose();
     (this._buffer as any) = null;
-    this._transparentMat.dispose();
-    this._meshColorMap.clear();
-    for (const id in this.materialCache) {
-      const material = this.materialCache.get(id);
-      if (material) {
-        material.dispose();
-      }
-    }
-    const disposer = this.components.tools.get(Disposer);
-    for (const id in this._colorMeshes) {
-      const mesh = this._colorMeshes.get(id);
-      if (mesh) {
-        disposer.destroy(mesh);
-      }
-    }
-    this._colorMeshes.clear();
-    this._meshes.clear();
     this.onDisposed.reset();
-  }
-
-  /**
-   * Adds a new mesh to be processed and managed by the culler.
-   * @mesh the mesh or instanced mesh to add.
-   */
-  add(mesh: THREE.Mesh | THREE.InstancedMesh) {
-    if (!this.enabled) return;
-
-    const isInstanced = mesh instanceof THREE.InstancedMesh;
-
-    const { geometry, material } = mesh;
-
-    const { r, g, b, code } = this.getNextColor();
-    const colorMaterial = this.getMaterial(r, g, b);
-
-    let newMaterial: THREE.Material[] | THREE.Material;
-
-    if (Array.isArray(material)) {
-      let transparentOnly = true;
-      const matArray: any[] = [];
-
-      for (const mat of material) {
-        if (this.isTransparent(mat)) {
-          matArray.push(this._transparentMat);
-        } else {
-          transparentOnly = false;
-          matArray.push(colorMaterial);
-        }
-      }
-
-      // If we find that all the materials are transparent then we must remove this from analysis
-      if (transparentOnly) {
-        colorMaterial.dispose();
-        return;
-      }
-
-      newMaterial = matArray;
-    } else if (this.isTransparent(material)) {
-      // This material is transparent, so we must remove it from analysis
-      colorMaterial.dispose();
-      return;
-    } else {
-      newMaterial = colorMaterial;
-    }
-
-    this._meshColorMap.set(code, mesh);
-
-    const count = isInstanced ? mesh.count : 1;
-    const colorMesh = new THREE.InstancedMesh(geometry, newMaterial, count);
-
-    if (isInstanced) {
-      colorMesh.instanceMatrix = mesh.instanceMatrix;
-    } else {
-      colorMesh.setMatrixAt(0, new THREE.Matrix4());
-    }
-
-    mesh.visible = false;
-
-    colorMesh.applyMatrix4(mesh.matrix);
-    colorMesh.updateMatrix();
-
-    this._scene.add(colorMesh);
-    this._colorMeshes.set(mesh.uuid, colorMesh);
-    this._meshes.set(mesh.uuid, mesh);
   }
 
   /**
@@ -229,7 +123,7 @@ export class CullerRenderer extends Component<
 
     this.renderer.setSize(this._width, this._height);
     this.renderer.setRenderTarget(this.renderTarget);
-    this.renderer.render(this._scene, camera);
+    this.renderer.render(this.scene, camera);
 
     const context = this.renderer.getContext() as WebGL2RenderingContext;
     await readPixelsAsync(
@@ -246,7 +140,7 @@ export class CullerRenderer extends Component<
     this.renderer.setRenderTarget(null);
 
     if (this.renderDebugFrame) {
-      this.renderer.render(this._scene, camera);
+      this.renderer.render(this.scene, camera);
     }
 
     this.worker.postMessage({
@@ -256,57 +150,7 @@ export class CullerRenderer extends Component<
     this.needsUpdate = false;
   };
 
-  private handleWorkerMessage = async (event: MessageEvent) => {
-    const colors = event.data.colors as Set<string>;
-
-    this._recentlyHiddenMeshes = new Set(this._currentVisibleMeshes);
-    this._currentVisibleMeshes.clear();
-
-    for (const code of colors.values()) {
-      const mesh = this._meshColorMap.get(code);
-      if (mesh) {
-        // mesh.visible = true;
-        this._currentVisibleMeshes.add(mesh);
-        this._recentlyHiddenMeshes.delete(mesh);
-      }
-    }
-
-    // for (const uuid of this._recentlyHiddenMeshes) {
-    //   const mesh = this._meshes.get(uuid);
-    //   if (mesh === undefined) continue;
-    //   mesh.visible = false;
-    // }
-
-    await this.onViewUpdated.trigger({
-      seen: this._currentVisibleMeshes,
-      unseen: this._recentlyHiddenMeshes,
-    });
-  };
-
-  private getMaterial(r: number, g: number, b: number) {
-    const colorEnabled = THREE.ColorManagement.enabled;
-    THREE.ColorManagement.enabled = false;
-    const code = `rgb(${r}, ${g}, ${b})`;
-    const color = new THREE.Color(code);
-    let material = this.materialCache.get(code);
-    const clippingPlanes = this.components.renderer.clippingPlanes;
-    if (!material) {
-      material = new THREE.MeshBasicMaterial({
-        color,
-        clippingPlanes,
-        side: THREE.DoubleSide,
-      });
-      this.materialCache.set(code, material);
-    }
-    THREE.ColorManagement.enabled = colorEnabled;
-    return material;
-  }
-
-  private isTransparent(material: THREE.Material) {
-    return material.transparent && material.opacity < 1;
-  }
-
-  private getNextColor() {
+  protected getNextColor() {
     if (this._colors.i === 0) {
       this._colors.b++;
       if (this._colors.b === 256) {
@@ -344,7 +188,7 @@ export class CullerRenderer extends Component<
   private applySettings(settings?: CullerRendererSettings) {
     if (settings) {
       if (settings.updateInterval !== undefined) {
-        this._updateInterval = settings.updateInterval;
+        this.updateInterval = settings.updateInterval;
       }
       if (settings.height !== undefined) {
         this._height = settings.height;
@@ -353,7 +197,7 @@ export class CullerRenderer extends Component<
         this._width = settings.width;
       }
       if (settings.autoUpdate !== undefined) {
-        this._autoUpdate = settings.autoUpdate;
+        this.autoUpdate = settings.autoUpdate;
       }
     }
   }
